Added staging data and minor updates.
TODO: Staging currently does not run everything on parallel. The
MC is created in parallel, but the 0v and data is done in serial.

TODO: Not happy with `create_staging.py`, it needs to be more
modular so we don't need to repeat code for different runs (mp vs no-mp)

TODO: Decide on the output columns list.

TODO: Decide where the cuts go.